requirements_cpu.txt
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keras2onnx==1.5.0
matplotlib==3.1.0
numpy==1.16.4
onnx==1.5.0
onnx-tf==1.3.0
onnxruntime==0.5.0
opencv-python==4.1.0.25
pandas==0.24.2
Pillow==6.0.0
PyGithub==1.43.8
scikit-image==0.15.0
scipy==1.3.0
seaborn==0.9.0
six==1.12.0
tensorboard==1.14.0
tensorflow==1.14.0
tf2onnx==1.5.1
torch==1.2.0
torchvision==0.4.0
tqdm==4.32.2
